+++ tests/core/txpool-module/test_txpool_content.py
@@ -8,7 +8,7 @@ from web3._utils.threads import (
 def test_txpool_content(web3_empty):
     web3 = web3_empty
 
-    web3.miner.stop()
+    web3.geth.miner.stop()
 
     with Timeout(60) as timeout:
         while web3.eth.hashrate or web3.eth.mining:
@@ -27,7 +27,7 @@ def test_txpool_content(web3_empty):
     })
     txn_2 = web3.eth.getTransaction(txn_2_hash)
 
-    content = web3.txpool.content
+    content = web3.geth.txpool.content
 
     assert web3.eth.coinbase in content['pending']