+++ tests/integration/go_ethereum/common.py
@@ -1,9 +1,10 @@
 import pytest
 
-from web3._utils.module_testing import (
+from web3._utils.module_testing import (  # noqa: F401
     EthModuleTest,
+    GoEthereumPersonalModuleTest,
+    GoEthereumShhModuleTest,
     NetModuleTest,
-    PersonalModuleTest,
     VersionModuleTest,
     Web3ModuleTest,
 )
@@ -59,6 +60,17 @@ class GoEthereumEthModuleTest(EthModuleTest):
             web3, unlocked_account_dual_type
         )
 
+    def test_eth_submitHashrate(self, web3):
+        if 'v1.8.22' in web3.clientVersion:
+            # https://github.com/ethereum/go-ethereum/commit/51db5975cc5fb88db6a0dba1826b534fd4df29d7
+            pytest.xfail('eth_submitHashrate deprecated in 1.8.22 for ethash_submitHashRate')
+        super().test_eth_submitHashrate(web3)
+
+    def test_eth_chainId(self, web3):
+        if 'v1.7.2' in web3.clientVersion:
+            pytest.xfail('eth_chainId not implemented in geth 1.7.2')
+        super().test_eth_chainId(web3)
+
 
 class GoEthereumVersionModuleTest(VersionModuleTest):
     pass
@@ -68,5 +80,18 @@ class GoEthereumNetModuleTest(NetModuleTest):
     pass
 
 
-class GoEthereumPersonalModuleTest(PersonalModuleTest):
-    pass
+class CommonGoEthereumShhModuleTest(GoEthereumShhModuleTest):
+    def test_shh_sync_filter(self, web3):
+        if 'v1.7.2' in web3.clientVersion:
+            pytest.xfail('Whisper version 6 not supported in geth 1.7.2')
+        super().test_shh_sync_filter(web3)
+
+    def test_shh_async_filter(self, web3):
+        if 'v1.7.2' in web3.clientVersion:
+            pytest.xfail('Whisper version 6 not supported in geth 1.7.2')
+        super().test_shh_async_filter(web3)
+
+    def test_shh_post(self, web3):
+        if 'v1.7.2' in web3.clientVersion:
+            pytest.xfail('Whisper version 6 not supported in geth 1.7.2')
+        super().test_shh_post(web3)