+++ tests/integration/parity/test_parity_ws.py
@@ -15,6 +15,7 @@ from web3._utils.module_testing import (
 )
 
 from .common import (
+    CommonParityShhModuleTest,
     ParityEthModuleTest,
     ParityPersonalModuleTest,
     ParityTraceModuleTest,
@@ -49,8 +50,10 @@ def parity_command_arguments(
         '--password', passwordfile,
         '--ws-port', ws_port,
         '--ws-origins', '*',
+        '--ws-apis', 'all',
         '--no-ipc',
         '--no-jsonrpc',
+        '--whisper',
     )
 
 
@@ -64,6 +67,7 @@ def parity_import_blocks_command(parity_binary, ws_port, datadir, passwordfile):
         '--password', passwordfile,
         '--ws-port', str(ws_port),
         '--ws-origins', '*',
+        '--ws-apis', 'all',
         '--no-ipc',
         '--no-jsonrpc',
         '--tracing', 'on',
@@ -102,4 +106,8 @@ class TestParityTraceModuleTest(ParityTraceModuleTest):
 
 
 class TestMiscWebsocketTest(MiscWebsocketTest):
+    pass
+
+
+class TestParityShhModuleTest(CommonParityShhModuleTest):
     pass