+++ web3/_utils/module_testing/__init__.py
@@ -8,12 +8,16 @@ from .net_module import (  # noqa: F401
     NetModuleTest,
 )
 from .personal_module import (  # noqa: F401
-    PersonalModuleTest,
+    GoEthereumPersonalModuleTest,
+    ParityPersonalModuleTest,
+)
+from .shh_module import (  # noqa: F401
+    GoEthereumShhModuleTest,
+    ParityShhModuleTest,
 )
 from .version_module import (  # noqa: F401
     VersionModuleTest,
 )
-
 from .parity_module import (  # noqa: F401
     ParityModuleTest,
 )