+++ web3/_utils/rpc_abi.py
@@ -50,7 +50,10 @@ RPC_ABIS = {
     'eth_newFilter': FILTER_PARAMS_ABIS,
     'eth_sendRawTransaction': ['bytes'],
     'eth_sendTransaction': TRANSACTION_PARAMS_ABIS,
+    'eth_signTransaction': TRANSACTION_PARAMS_ABIS,
     'eth_sign': ['address', 'bytes'],
+    'eth_submitHashrate': ['uint', 'bytes32'],
+    'eth_submitWork': ['bytes8', 'bytes32', 'bytes32'],
     # personal
     'personal_sendTransaction': TRANSACTION_PARAMS_ABIS,
     'personal_lockAccount': ['address'],