+++ web3/auto/infura/endpoints.py
@@ -1,12 +1,16 @@
-import logging
 import os
 
 from eth_utils import (
     ValidationError,
 )
 
+from web3.exceptions import (
+    InfuraKeyNotFound,
+)
+
 INFURA_MAINNET_DOMAIN = 'mainnet.infura.io'
 INFURA_ROPSTEN_DOMAIN = 'ropsten.infura.io'
+INFURA_GOERLI_DOMAIN = 'goerli.infura.io'
 INFURA_RINKEBY_DOMAIN = 'rinkeby.infura.io'
 INFURA_KOVAN_DOMAIN = 'kovan.infura.io'
 
@@ -19,25 +23,32 @@ def load_api_key():
     key = os.environ.get('WEB3_INFURA_PROJECT_ID',
                          os.environ.get('WEB3_INFURA_API_KEY', ''))
     if key == '':
-        logging.getLogger('web3.auto.infura').warning(
-            "No Infura Project ID found. Add environment variable WEB3_INFURA_PROJECT_ID to "
-            " ensure continued API access after March 27th. "
-            "New keys are available at https://infura.io/register"
+        raise InfuraKeyNotFound(
+            "No Infura Project ID found. Please ensure "
+            "that the environment variable WEB3_INFURA_PROJECT_ID is set."
         )
     return key
+
+
+def load_secret():
+    return os.environ.get('WEB3_INFURA_API_SECRET', '')
+
+
+def build_http_headers():
+    secret = load_secret()
+    if secret:
+        headers = {'auth': ('', secret)}
+        return headers
 
 
 def build_infura_url(domain):
     scheme = os.environ.get('WEB3_INFURA_SCHEME', WEBSOCKET_SCHEME)
     key = load_api_key()
+    secret = load_secret()
 
-    if key and scheme == WEBSOCKET_SCHEME:
-        return "%s://%s/ws/v3/%s" % (scheme, domain, key)
-    elif key and scheme == HTTP_SCHEME:
-        return "%s://%s/v3/%s" % (scheme, domain, key)
-    elif scheme == WEBSOCKET_SCHEME:
-        return "%s://%s/ws/" % (scheme, domain)
+    if scheme == WEBSOCKET_SCHEME:
+        return "%s://:%s@%s/ws/v3/%s" % (scheme, secret, domain, key)
     elif scheme == HTTP_SCHEME:
-        return "%s://%s/%s" % (scheme, domain, key)
+        return "%s://%s/v3/%s" % (scheme, domain, key)
     else:
         raise ValidationError("Cannot connect to Infura with scheme %r" % scheme)