+++ web3/eth.py
@@ -13,9 +13,6 @@ from hexbytes import (
 from web3._utils.blocks import (
     select_method_for_block_identifier,
 )
-from web3._utils.decorators import (
-    deprecated_for,
-)
 from web3._utils.empty import (
     empty,
 )
@@ -46,7 +43,9 @@ from web3.contract import (
     Contract,
 )
 from web3.exceptions import (
+    BlockNotFound,
     TimeExhausted,
+    TransactionNotFound,
 )
 from web3.iban import (
     Iban,
@@ -63,10 +62,6 @@ class Eth(Module):
     defaultContractFactory = Contract
     iban = Iban
     gasPriceStrategy = None
-
-    @deprecated_for("doing nothing at all")
-    def enable_unaudited_features(self):
-        pass
 
     def namereg(self):
         raise NotImplementedError()
@@ -106,6 +101,10 @@ class Eth(Module):
     def blockNumber(self):
         return self.web3.manager.request_blocking("eth_blockNumber", [])
 
+    @property
+    def chainId(self):
+        return self.web3.manager.request_blocking("eth_chainId", [])
+
     def getBalance(self, account, block_identifier=None):
         if block_identifier is None:
             block_identifier = self.defaultBlock
@@ -142,10 +141,13 @@ class Eth(Module):
             if_number='eth_getBlockByNumber',
         )
 
-        return self.web3.manager.request_blocking(
+        result = self.web3.manager.request_blocking(
             method,
             [block_identifier, full_transactions],
         )
+        if result is None:
+            raise BlockNotFound(f"Block with id: {block_identifier} not found.")
+        return result
 
     def getBlockTransactionCount(self, block_identifier):
         """
@@ -158,10 +160,13 @@ class Eth(Module):
             if_hash='eth_getBlockTransactionCountByHash',
             if_number='eth_getBlockTransactionCountByNumber',
         )
-        return self.web3.manager.request_blocking(
+        result = self.web3.manager.request_blocking(
             method,
             [block_identifier],
         )
+        if result is None:
+            raise BlockNotFound(f"Block with id: {block_identifier} not found.")
+        return result
 
     def getUncleCount(self, block_identifier):
         """
@@ -174,10 +179,13 @@ class Eth(Module):
             if_hash='eth_getUncleCountByBlockHash',
             if_number='eth_getUncleCountByBlockNumber',
         )
-        return self.web3.manager.request_blocking(
+        result = self.web3.manager.request_blocking(
             method,
             [block_identifier],
         )
+        if result is None:
+            raise BlockNotFound(f"Block with id: {block_identifier} not found.")
+        return result
 
     def getUncleByBlock(self, block_identifier, uncle_index):
         """
@@ -190,24 +198,31 @@ class Eth(Module):
             if_hash='eth_getUncleByBlockHashAndIndex',
             if_number='eth_getUncleByBlockNumberAndIndex',
         )
-        return self.web3.manager.request_blocking(
+        result = self.web3.manager.request_blocking(
             method,
             [block_identifier, uncle_index],
         )
+        if result is None:
+            raise BlockNotFound(
+                f"Uncle at index: {uncle_index} of block with id: {block_identifier} not found."
+            )
+        return result
 
     def getTransaction(self, transaction_hash):
-        return self.web3.manager.request_blocking(
+        result = self.web3.manager.request_blocking(
             "eth_getTransactionByHash",
             [transaction_hash],
         )
+        if result is None:
+            raise TransactionNotFound(f"Transaction with hash: {transaction_hash} not found.")
+        return result
 
-    @deprecated_for("w3.eth.getTransactionByBlock")
     def getTransactionFromBlock(self, block_identifier, transaction_index):
         """
         Alias for the method getTransactionByBlock
         Depreceated to maintain naming consistency with the json-rpc API
         """
-        return self.getTransactionByBlock(block_identifier, transaction_index)
+        raise DeprecationWarning("This method has been deprecated as of EIP 1474.")
 
     def getTransactionByBlock(self, block_identifier, transaction_index):
         """
@@ -220,10 +235,16 @@ class Eth(Module):
             if_hash='eth_getTransactionByBlockHashAndIndex',
             if_number='eth_getTransactionByBlockNumberAndIndex',
         )
-        return self.web3.manager.request_blocking(
+        result = self.web3.manager.request_blocking(
             method,
             [block_identifier, transaction_index],
         )
+        if result is None:
+            raise TransactionNotFound(
+                f"Transaction index: {transaction_index} "
+                f"on block id: {block_identifier} not found."
+            )
+        return result
 
     def waitForTransactionReceipt(self, transaction_hash, timeout=120):
         try:
@@ -237,20 +258,20 @@ class Eth(Module):
             )
 
     def getTransactionReceipt(self, transaction_hash):
-        return self.web3.manager.request_blocking(
+        result = self.web3.manager.request_blocking(
             "eth_getTransactionReceipt",
             [transaction_hash],
         )
+        if result is None:
+            raise TransactionNotFound(f"Transaction with hash: {transaction_hash} not found.")
+        return result
 
     def getTransactionCount(self, account, block_identifier=None):
         if block_identifier is None:
             block_identifier = self.defaultBlock
         return self.web3.manager.request_blocking(
             "eth_getTransactionCount",
-            [
-                account,
-                block_identifier,
-            ],
+            [account, block_identifier],
         )
 
     def replaceTransaction(self, transaction_hash, new_transaction):
@@ -292,6 +313,11 @@ class Eth(Module):
         message_hex = to_hex(data, hexstr=hexstr, text=text)
         return self.web3.manager.request_blocking(
             "eth_sign", [account, message_hex],
+        )
+
+    def signTransaction(self, transaction):
+        return self.web3.manager.request_blocking(
+            "eth_signTransaction", [transaction],
         )
 
     @apply_to_return_value(HexBytes)
@@ -373,6 +399,16 @@ class Eth(Module):
             "eth_getLogs", [filter_params],
         )
 
+    def submitHashrate(self, hashrate, node_id):
+        return self.web3.manager.request_blocking(
+            "eth_submitHashrate", [hashrate, node_id],
+        )
+
+    def submitWork(self, nonce, pow_hash, mix_digest):
+        return self.web3.manager.request_blocking(
+            "eth_submitWork", [nonce, pow_hash, mix_digest],
+        )
+
     def uninstallFilter(self, filter_id):
         return self.web3.manager.request_blocking(
             "eth_uninstallFilter", [filter_id],
@@ -394,7 +430,7 @@ class Eth(Module):
         self.defaultContractFactory = contractFactory
 
     def getCompilers(self):
-        return self.web3.manager.request_blocking("eth_getCompilers", [])
+        raise DeprecationWarning("This method has been deprecated as of EIP 1474.")
 
     def getWork(self):
         return self.web3.manager.request_blocking("eth_getWork", [])